Refactor comment timestamp display in Comments component
The code changes in this commit refactor the way comment timestamps are displayed in the Comments component. The previous implementation included both the hour and minute values, but they have been removed to simplify the display. Now, only the date and time in 12-hour format are shown. Additionally, a CSS class has been added to style the timestamp text as smaller (text-sm).